Setting it to Uint8 caused it to be treated as an unsigned char, which was problematic

+++ main/Color24.hpp
@@ -15,7 +15,7 @@
 struct Color24
 {
 private:
-	typedef Uint8 ColorType;
+	typedef unsigned short ColorType;
 
 public:
 	union
